ViteJS/Templating/SSR for poem
