+++ create-rust-app/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "create-rust-app"
 description = "Set up a modern rust+react web app by running one command."
-version = "6.0.2"
+version = "7.0.0"
 edition = "2018"
 authors = ["Haris <4259838+Wulf@users.noreply.github.com>"]
 readme = "../README.md"
@@ -11,16 +11,17 @@ keywords = ["react", "typescript", "generation", "backend", "frontend"]
 categories = ["command-line-utilities", "development-tools", "web-programming", "config", "database"]
 
 [dependencies]
+##
+## COMMON / DEFAULT - required dependencies
+##
 dotenv = "0.15.0"
-serde_json = "1.0.74"
+serde_json = "1.0.81"
 lettre = "0.9.6"
 lettre_email = "0.9.4"
-
-##
-## COMMON - required dependencies
-##
 tera = { version="1.15.0" }
 lazy_static = { version="1.4.0" }
+serde = { version = "1.0.137", features = ["derive"] }
+diesel = { version="1.4.8", default-features = false, features = ["postgres", "uuid", "r2d2", "chrono"] } # plugin_dev, plugin_auth
 
 ##
 ## PLUGINS
@@ -46,7 +47,6 @@ http = { optional=true, version="0.2.6" }
 diesel_derives = { optional=true, version="1.4.1" }
 uuid = { optional=true, version="0.8.2", features=["v4", "serde"] }
 md5 = { optional=true, version="0.7.0" }
-mime_guess = { optional=true, version="2.0.4" }
 base64 = { optional=true, version="0.13.0" }
 
 ##
@@ -54,14 +54,14 @@ base64 = { optional=true, version="0.13.0" }
 ##
 
 # poem dependencies
-poem = { optional=true, version="1.2.33", features=["anyhow", "cookie", "static-files"] }
+poem = { optional=true, version="1.3.30", features=["anyhow", "cookie", "static-files"] }
 
 # actix_web dependencies
 actix-multipart = { optional=true, version="0.4.0" }
 actix-files = { optional=true, version="0.6.0" }
-actix-http = { optional=true, version="3.0.0" }
+actix-http = { optional=true, version="3.0.4" }
 actix-web = { optional=true, version="4.0.1" }
-actix-web-httpauth = { optional=true, version="0.6.0-beta.8" }
+actix-web-httpauth = { optional=true, version="0.6.0" }
 derive_more = { optional=true, version="0.99.17" }
 futures = { optional=true, version="0.3.21" }
 env_logger = { optional=true, version= "0.9.0" }
@@ -70,16 +70,15 @@ env_logger = { optional=true, version= "0.9.0" }
 ## MISC - here, we list deps which are required by multiple features but are not required in all configurations
 ##
 
-serde = { optional=true, version = "1.0.133", features = ["derive"] } # plugin_dev, plugin_auth
-anyhow = { optional=true, version="1.0.52" } # backend_poem, plugin_auth
-diesel = { optional=true, version="1.4.8", default-features = false, features = ["postgres", "uuid", "r2d2", "chrono", "32-column-tables"] } # plugin_dev, plugin_auth
+mime_guess = { optional=true, version="2.0.4" } # backend_poem, plugin_storage
+anyhow = { optional=true, version="1.0.57" } # backend_poem, plugin_auth
 
 [features]
-default = ["plugin_dev", "plugin_auth", "plugin_container"]
-plugin_dev = ["diesel", "serde", "diesel_migrations"]
+default = []
+plugin_dev = ["diesel_migrations"]
 plugin_container = []
-plugin_auth = ["diesel", "serde", "anyhow", "argonautica", "jsonwebtoken", "chrono", "tsync"]
-plugin_storage = [ "aws-config", "aws-types", "aws-endpoint", "aws-sdk-s3", "tokio", "futures-util", "http", "diesel", "diesel_derives", "serde", "uuid", "md5", "mime_guess", "base64" ]
+plugin_auth = ["anyhow", "argonautica", "jsonwebtoken", "chrono", "tsync"]
+plugin_storage = [ "aws-config", "aws-types", "aws-endpoint", "aws-sdk-s3", "tokio", "futures-util", "http", "diesel_derives", "uuid", "md5", "mime_guess", "base64" ]
 plugin_graphql = []
-backend_poem = ["poem"]
+backend_poem = ["poem", "anyhow", "mime_guess"]
 backend_actix-web = ["actix-web", "actix-http", "actix-files", "actix-multipart", "actix-web-httpauth","derive_more", "futures", "env_logger"]
