+++ create-rust-app/src/auth/extractors/mod.rs
@@ -1,0 +1,9 @@
+#[cfg(feature = "backend_actix-web")]
+mod auth_actixweb;
+#[cfg(feature = "backend_actix-web")]
+pub use auth_actixweb::Auth;
+
+#[cfg(feature = "backend_poem")]
+mod auth_poem;
+#[cfg(feature = "backend_poem")]
+pub use auth_poem::Auth;
