+++ create-rust-app/src/database.rs
@@ -1,6 +1,3 @@
-// #[macro_use]
-// extern crate diesel;
-
 use diesel::{
     r2d2::{self, ConnectionManager, PooledConnection},
     PgConnection,
