+++ create-rust-app/src/dev/endpoints/service_actixweb.rs
@@ -1,0 +1,55 @@
+use std::ops::Deref;
+use actix_web::{get, HttpResponse, post, Scope, web::{Data, Json}};
+use crate::{auth::Auth, Database, dev::controller, dev::controller::{HealthCheckResponse, MyQueryResult, MySqlQuery}};
+
+#[post("/db/query")]
+async fn query_db(db: Data<Database>, body: Json<MySqlQuery>) -> HttpResponse {
+    match controller::query_db(&db, body.deref()) {
+        Ok(result) => HttpResponse::Ok().body(result),
+        Err(t) => HttpResponse::InternalServerError().finish()
+    }
+}
+
+#[get("/auth/has-system-role")]
+async fn check_system_role(auth: Auth) -> HttpResponse {
+    HttpResponse::Ok().json(controller::check_system_role(&auth))
+}
+
+#[get("/auth/add-system-role")]
+async fn add_system_role(db: Data<Database>, auth: Auth) -> HttpResponse {
+    HttpResponse::Ok().json(controller::add_system_role(&db, &auth))
+}
+
+#[get("/db/is-connected")]
+async fn is_connected(db: Data<Database>) -> HttpResponse {
+    HttpResponse::Ok().json(controller::is_connected(&db))
+}
+
+#[get("/db/needs-migration")]
+async fn needs_migration(db: Data<Database>) -> HttpResponse {
+    HttpResponse::Ok().json(controller::needs_migration(&db))
+}
+
+#[get("/db/migrate")]
+async fn migrate_db(db: Data<Database>) -> HttpResponse {
+    HttpResponse::Ok().json(controller::migrate_db(&db))
+}
+
+#[get("/health")]
+async fn health() -> HttpResponse {
+    controller::health();
+    HttpResponse::Ok().json(HealthCheckResponse {
+        message: "healthy".to_string(),
+    })
+}
+
+pub fn endpoints(scope: Scope) -> Scope {
+    scope
+        .service(query_db)
+        .service(check_system_role)
+        .service(add_system_role)
+        .service(is_connected)
+        .service(needs_migration)
+        .service(migrate_db)
+        .service(health)
+}
