+++ create-rust-app/src/util/mod.rs
@@ -3,12 +3,19 @@
 /// are exposed directly as create_rust_app::<utilty-fn>.
 ///
 
-
 #[cfg(feature = "backend_actix-web")]
 mod actix_web_utils;
 
 #[cfg(feature = "backend_actix-web")]
 pub use actix_web_utils::*;
 
+#[cfg(feature = "backend_poem")]
+mod poem_utils;
+
+#[cfg(feature = "backend_poem")]
+pub use poem_utils::*;
+
 /// expose template_utils for all backends.
 mod template_utils;
+
+pub mod net;
