+++ create-rust-app_cli/src/content/project.rs
@@ -40,7 +40,7 @@ struct Asset;
 //     version.to_string()
 // }
 fn get_current_cra_lib_version() -> String {
-    "6.0.2".to_string()
+    "7.0.0".to_string()
 }
 
 fn add_bins_to_cargo_toml(project_dir: &std::path::PathBuf) -> Result<(), std::io::Error> {
