+++ create-rust-app_cli/template-plugin-dev/.cargo/admin/.proxyrc
@@ -1,5 +1,5 @@
 {
   "/api": {
-    "target": "http://localhost:8080",
+    "target": "http://localhost:3000",
   }
 }
