+++ create-rust-app_cli/template/backend/main.rs+actix_web
@@ -35,5 +35,5 @@ async fn main() -> std::io::Result<()> {
         app = app.service(api_scope);
         app = app.default_service(web::get().to(create_rust_app::render_views));
         app
-    }).bind("0.0.0.0:8080")?.run().await
+    }).bind("0.0.0.0:3000")?.run().await
 }
